separate game class
